fix(deployments): use correct FilterValue::Integer variant
Replace FilterValue::BigInt (non-existent) with FilterValue::Integer
for cluster_id validation query.

+++ app/src/apps/deployments/views/create_deployment.rs
@@ -33,7 +33,7 @@ pub async fn create_deployment(request: Request) -> ViewResult<Response> {
 		.filter(
 			Cluster::field_id(),
 			FilterOperator::Eq,
-			FilterValue::BigInt(body.cluster_id),
+			FilterValue::Integer(body.cluster_id),
 		)
 		.first()
 		.await
